+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Addr:full tag is not set when addr:housenumber is present.
 * Whitespace is stripped from tag values in a dataset.
+* Conflate `contact:` namespace.
 
 ## 1.2.1